:bug: fix: authentication options to seperate lib file
