+++ src/app/progress/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import MetricsSection from './Metrics';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/next-auth/authOptions';
 import { redirect } from 'next/navigation';
 
 // Sample data structure remains the same as before
